pull use xmtp client out to own hook
